+++ src/components/elements/upload/Uploader.stories.tsx
@@ -3,7 +3,7 @@ import { Box } from '@mui/system';
 import { Meta, StoryObj } from '@storybook/react';
 
 import MOCK_IMAGE from './MOCK_IMAGE';
-import Uploader from './UploaderBase';
+import Uploader from './Uploader';
 
 import { createDirectUploadMock, getFileMock } from '@/test/__mocks__/requests';
