+++ src/components/elements/upload/Uploader.tsx
@@ -17,14 +17,11 @@ import {
   DropzoneRootProps,
   useDropzone,
 } from 'react-dropzone';
-
-import {
-  CurrentFileSummary,
-  ExistingFileSummary,
-  FilePreviewIcon,
-} from './FileSummary';
 import useDirectUpload from './useDirectUpload';
 
+import FilePreviewIcon from '@/components/elements/upload/FilePreviewIcon';
+import SavedFileSummary from '@/components/elements/upload/SavedFileSummary';
+import UploadedFileSummary from '@/components/elements/upload/UploadedFileSummary';
 import { DirectUpload, FileFieldsFragment } from '@/types/gqlTypes';
 import { ensureArray } from '@/utils/arrays';
 
@@ -271,14 +268,14 @@ const Uploader: React.FC<UploaderProps> = ({
               </>
             )}
           {!loading && !multiple && currentFiles[0] && (
-            <CurrentFileSummary
+            <UploadedFileSummary
               file={currentFiles[0]}
               onRemove={() => removeFile(currentFiles[0])}
               variant='stacked'
             />
           )}
           {!loading && !multiple && existingFiles[0] && (
-            <ExistingFileSummary
+            <SavedFileSummary
               file={existingFiles[0] as FileFieldsFragment}
               onRemove={() =>
                 removeFile(existingFiles[0] as FileFieldsFragment)
@@ -303,7 +300,7 @@ const Uploader: React.FC<UploaderProps> = ({
           <Stack divider={<Divider />}>
             {existingFiles.map((file) => {
               return (
-                <ExistingFileSummary
+                <SavedFileSummary
                   key={(file as FileFieldsFragment).id} // it will always be a FileFieldsFragment, ts is just confused
                   file={file as FileFieldsFragment}
                   onRemove={() => removeFile(file as FileFieldsFragment)}
@@ -313,7 +310,7 @@ const Uploader: React.FC<UploaderProps> = ({
             })}
             {currentFiles.map((file) => {
               return (
-                <CurrentFileSummary
+                <UploadedFileSummary
                   key={file.name} // we enforce uniqueness on file names
                   file={file}
                   onRemove={() => removeFile(file)}
