+++ src/modules/form/components/DynamicField.tsx
@@ -43,7 +43,7 @@ import TextInput from '@/components/elements/input/TextInput';
 import TimeOfDayPicker from '@/components/elements/input/TimeOfDayPicker';
 import YesNoRadio from '@/components/elements/input/YesNoRadio';
 import LabelWithContent from '@/components/elements/LabelWithContent';
-import Uploader from '@/components/elements/upload/UploaderBase';
+import Uploader from '@/components/elements/upload/Uploader';
 import useAuth from '@/modules/auth/hooks/useAuth';
 import MciClearance from '@/modules/external/mci/components/MciClearance';
 import SimpleAddressInput from '@/modules/form/components/client/addresses/SimpleAddressInput';
